Temporary fix to the build
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@601712 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/render/afp/AFPRenderer.java
@@ -600,7 +600,7 @@ public class AFPRenderer extends AbstractPathOrientedRenderer {
             AFPFontAttributes afpFontAttributes = (AFPFontAttributes) i.next();
 
             afpDataStream.createFont(
-                afpFontAttributes.getFontReference(),
+                (byte)afpFontAttributes.getFontReference(),
                 afpFontAttributes.getFont(),
                 afpFontAttributes.getPointSize());
